Finishing Devolution Screens

+++ demo/src/main/java/com/library/controller/BookController.java
@@ -1,9 +1,15 @@
 package com.library.controller;
 
 import com.library.exceptions.BookNotFoundException;
+import com.library.exceptions.UnableToDeleteBook;
 import com.library.exceptions.UninformedParameterException;
+import com.library.exceptions.UserNotFoundException;
 import com.library.model.entities.Book;
+import com.library.model.entities.Loan;
+import com.library.model.entities.LoanStatus;
+import com.library.model.entities.User;
 import com.library.model.service.BookService;
+import com.library.model.service.LoanService;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -31,7 +37,10 @@ public class BookController {
         bookService.updateBook(book);
     }
 
-    public void deleteBook(Integer id) {
+    public void deleteBook(Integer id) throws BookNotFoundException, UnableToDeleteBook {
+        if (thisBookHasALoan(id)) {
+            throw new UnableToDeleteBook();
+        }
         bookService.deleteBookById(id);
     }
 
@@ -46,5 +55,17 @@ public class BookController {
 
     public List<Book> getBooks() {
         return bookService.getAllBooks();
+    }
+
+    private boolean thisBookHasALoan(Integer id) throws BookNotFoundException {
+        Book book = getBook(id);
+
+        if (book != null) {
+            LoanService loanService = new LoanService();
+            List<Loan> loans = loanService.getLoanByBook(book).stream().filter(loan -> loan.getLoanStatus() == LoanStatus.NOT_RETURNED).toList();
+
+            return !loans.isEmpty();
+        }
+        return false;
     }
 }
